+++ js/models/index.js
@@ -11,5 +11,6 @@ export { AudioElement } from './AudioElement.js';
 export { ShapeElement } from './ShapeElement.js';
 export { ListElement } from './ListElement.js';
 export { LinkElement } from './LinkElement.js';
+export { CountdownTimerElement } from './CountdownTimerElement.js';
 export { Slide } from './Slide.js';
 export { Presentation } from './Presentation.js';